Refactor lint issues in validation
