+++ src/queryAnalysis/validation/sql_server_login_test.go
@@ -2,83 +2,36 @@ package validation
 
 import (
 	"fmt"
-	"testing"
-
-	"github.com/jmoiron/sqlx"
-	"github.com/newrelic/nri-mssql/src/queryAnalysis/connection"
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/DATA-DOG/go-sqlmock.v1"
+	"testing"
 )
 
-func TestCheckSQLServerLoginEnabled(t *testing.T) {
-	// Create a new mock database connection
-	db, mock, err := sqlmock.New()
-	assert.NoError(t, err)
-	defer db.Close()
+func TestCheckPermissionsAndLogin_LoginEnabled(t *testing.T) {
+	sqlConnection, mock := setupMockDB(t)
+	defer sqlConnection.Connection.Close()
 
-	sqlConnection := &connection.SQLConnection{Connection: sqlx.NewDb(db, "sqlmock")}
+	mockCheckPermissions(mock, true)
+	mockCheckSQLServerLoginEnabled(mock, true)
 
-	// Define the expected rows
-	rows := sqlmock.NewRows([]string{"is_login_enabled"}).AddRow(true)
-
-	mock.ExpectQuery("SELECT CASE WHEN SERVERPROPERTY\\('IsIntegratedSecurityOnly'\\) = 0 THEN 1 ELSE 0 END AS is_login_enabled").
-		WillReturnRows(rows)
-
-	// Call the function
-	isLoginEnabled, err := checkSQLServerLoginEnabled(sqlConnection)
-
-	// Assertions
-	assert.NoError(t, err)
-	assert.True(t, isLoginEnabled)
-
-	// Ensure all expectations were met
+	result := checkPermissionsAndLogin(sqlConnection)
+	assert.True(t, result)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
 
-func TestCheckSQLServerLoginEnabled_Disabled(t *testing.T) {
-	// Create a new mock database connection
-	db, mock, err := sqlmock.New()
-	assert.NoError(t, err)
-	defer db.Close()
+func TestCheckPermissionsAndLogin_LoginEnabledError(t *testing.T) {
+	sqlConnection, mock := setupMockDB(t)
+	defer sqlConnection.Connection.Close()
 
-	sqlConnection := &connection.SQLConnection{Connection: sqlx.NewDb(db, "sqlmock")}
+	// Mock checkPermissions
+	mock.ExpectQuery("SELECT CASE WHEN IS_SRVROLEMEMBER\\('sysadmin'\\) = 1 OR HAS_PERMS_BY_NAME\\(null, null, 'VIEW SERVER STATE'\\) = 1 THEN 1 ELSE 0 END AS has_permission").
+		WillReturnRows(sqlmock.NewRows([]string{"has_permission"}).AddRow(true))
 
-	// Define the expected rows
-	rows := sqlmock.NewRows([]string{"is_login_enabled"}).AddRow(false)
-
-	mock.ExpectQuery("SELECT CASE WHEN SERVERPROPERTY\\('IsIntegratedSecurityOnly'\\) = 0 THEN 1 ELSE 0 END AS is_login_enabled").
-		WillReturnRows(rows)
-
-	// Call the function
-	isLoginEnabled, err := checkSQLServerLoginEnabled(sqlConnection)
-
-	// Assertions
-	assert.NoError(t, err)
-	assert.False(t, isLoginEnabled)
-
-	// Ensure all expectations were met
-	assert.NoError(t, mock.ExpectationsWereMet())
-}
-
-func TestCheckSQLServerLoginEnabled_Error(t *testing.T) {
-	// Create a new mock database connection
-	db, mock, err := sqlmock.New()
-	assert.NoError(t, err)
-	defer db.Close()
-
-	sqlConnection := &connection.SQLConnection{Connection: sqlx.NewDb(db, "sqlmock")}
-
-	// Define the expected error
+	// Mock checkSQLServerLoginEnabled error
 	mock.ExpectQuery("SELECT CASE WHEN SERVERPROPERTY\\('IsIntegratedSecurityOnly'\\) = 0 THEN 1 ELSE 0 END AS is_login_enabled").
 		WillReturnError(fmt.Errorf("query error"))
 
-	// Call the function
-	isLoginEnabled, err := checkSQLServerLoginEnabled(sqlConnection)
-
-	// Assertions
-	assert.Error(t, err)
-	assert.False(t, isLoginEnabled)
-
-	// Ensure all expectations were met
+	result := checkPermissionsAndLogin(sqlConnection)
+	assert.False(t, result)
 	assert.NoError(t, mock.ExpectationsWereMet())
 }
